+++ packages/react/src/components/Table/Table.main.story.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 import { object, select, boolean, text, number } from '@storybook/addon-knobs';
-import { merge, uniqueId } from 'lodash-es';
+import { cloneDeep, debounce, merge, uniqueId } from 'lodash-es';
+import { ToastNotification } from 'carbon-components-react';
 import { SettingsAdjust16 } from '@carbon/icons-react';
 
 import StoryNotice from '../../internal/StoryNotice';
@@ -24,6 +25,7 @@ import SearchingREADME from './mdx/Searching.mdx';
 import StatesREADME from './mdx/States.mdx';
 import PaginationREADME from './mdx/Pagination.mdx';
 import ToolbarREADME from './mdx/Toolbar.mdx';
+import EditDataREADME from './mdx/EditData.mdx';
 import Table from './Table';
 import StatefulTable from './StatefulTable';
 import {
@@ -56,6 +58,9 @@ import {
   getCustomToolbarContentElement,
 } from './Table.story.helpers';
 
+// Dataset used to speed up stories using row edit
+const storyTableData = getTableData();
+
 export default {
   title: '1 - Watson IoT/Table',
   parameters: {
@@ -80,7 +85,7 @@ export const Playground = () => {
   const [rowActionsState, setRowActionsState] = useStoryState(getRowActionStates());
 
   // KNOBS
-  // The order of appearance is defined function getTableKnobs.
+  // The order of appearance is defined by the function getTableKnobs.
   const {
     selectedTableType,
     tableMaxWidth,
@@ -212,7 +217,7 @@ export const Playground = () => {
 
   // INITIAL DATA STATE
   const [data, setData] = useStoryState(
-    [...(demoEmptyState || demoCustomEmptyState ? [] : getTableData())]
+    [...(demoEmptyState || demoCustomEmptyState ? [] : storyTableData)]
       .slice(0, numberOfRows)
       .map((row, index) => (hasRowActions ? addRowAction(row, hasSingleRowEdit, index) : row))
       .map((row, index) => (hasRowNesting ? addChildRows(row, index) : row))
@@ -1136,5 +1141,200 @@ WithToolbar.parameters = {
   component: Table,
   docs: {
     page: ToolbarREADME,
+  },
+};
+
+export const WithDataEditing = () => {
+  const {
+    selectedTableType,
+    hasRowActions,
+    hasRowEdit,
+    hasSingleRowEdit,
+    preserveCellWhiteSpace,
+  } = getTableKnobs({
+    knobsToCreate: [
+      'selectedTableType',
+      'hasRowActions',
+      'hasRowEdit',
+      'hasSingleRowEdit',
+      'preserveCellWhiteSpace',
+    ],
+    getDefaultValue: (knobName) => (knobName === 'selectedTableType' ? 'Table' : true),
+  });
+
+  const MyTable = selectedTableType === 'StatefulTable' ? StatefulTable : Table;
+  const editAction = getOverflowEditRowAction();
+  editAction.disabled = false;
+  const initialData = storyTableData.slice(0, 10).map((i) => ({
+    ...i,
+    rowActions: [editAction],
+  }));
+
+  const [showRowEditBar, setShowRowEditBar] = useStoryState(false);
+  const [currentData, setCurrentData] = useStoryState(initialData);
+  const [rowEditData, setRowEditedData] = useStoryState([]);
+  const [previousData, setPreviousData] = useStoryState([]);
+  const [showToast, setShowToast] = useStoryState(false);
+  const [rowActionsState, setRowActionsState] = useStoryState([]);
+  const [isPristine, setIsPristine] = useStoryState(true);
+
+  const disableRowActions = (data, disabled) => {
+    const rowAction = getOverflowEditRowAction();
+    rowAction.disabled = disabled;
+    return data.map((row) => ({ ...row, rowActions: [rowAction] }));
+  };
+
+  const onDataChange = debounce((newValue, columnId, rowId) => {
+    setRowEditedData((previousData) =>
+      previousData.map((row) =>
+        row.id === rowId ? { ...row, values: { ...row.values, [columnId]: newValue } } : row
+      )
+    );
+    setIsPristine(false);
+  });
+
+  const onShowMultiRowEdit = () => {
+    setRowEditedData(cloneDeep(currentData));
+    setShowRowEditBar(true);
+    setShowToast(false);
+    setCurrentData((prev) => disableRowActions(prev, true));
+  };
+  const onCancelRowEdit = () => {
+    setRowEditedData([]);
+    setShowRowEditBar(false);
+    setRowActionsState([]);
+    setIsPristine(true);
+    setCurrentData((prev) => disableRowActions(prev, false));
+  };
+  const onSaveRowEdit = () => {
+    // because of the nature of rendering these buttons dynamically (and asyncronously via dispatch)
+    // in the StatefulTable we need to wrap these calls inside the setRowEditedData callback to ensure
+    // we're always working with the correctly updated data.
+    setRowEditedData((prev) => {
+      setShowToast(true);
+      setPreviousData(disableRowActions(currentData, false));
+      setCurrentData(disableRowActions(prev, false));
+      setShowRowEditBar(false);
+      setRowActionsState([]);
+      setIsPristine(true);
+      return [];
+    });
+  };
+  const onUndoRowEdit = () => {
+    setCurrentData(previousData);
+    setPreviousData([]);
+    setShowToast(false);
+  };
+
+  const onApplyRowAction = (action, rowId) => {
+    if (action === 'edit') {
+      setRowEditedData(cloneDeep(currentData));
+      setCurrentData(disableRowActions(currentData, true));
+      setRowActionsState([{ rowId, isEditMode: true }]);
+    }
+  };
+
+  const saveCancelButtons = (
+    <div style={{ display: 'flex', gap: '0.5rem' }}>
+      <Button key="cancel" size="small" kind="tertiary" onClick={onCancelRowEdit}>
+        Cancel
+      </Button>
+      <Button key="save" size="small" onClick={onSaveRowEdit} disabled={isPristine}>
+        Save
+      </Button>
+    </div>
+  );
+
+  // We define some initial column widths to prevent the columns from adjusting
+  // in size when edit mode is activated
+  const columnWidths = {
+    string: '250px',
+    date: '220px',
+    select: '160px',
+    secretField: '165px',
+    status: '165px',
+    number: '95px',
+    boolean: '90px',
+    node: '115px',
+    object: '110px',
+  };
+  const columns = getTableColumns().map((column) => ({
+    ...column,
+    // This is a simplified example.
+    // The app should also handle input validation etc
+    editDataFunction: getEditDataFunction(onDataChange),
+    width: columnWidths[column.id],
+  }));
+
+  const myToast = (
+    <ToastNotification
+      style={{ position: 'absolute', zIndex: '99' }}
+      hideCloseButton={false}
+      kind="success"
+      notificationType="inline"
+      role="alert"
+      subtitle={
+        <div style={{ display: 'flex', alignItems: 'center', marginTop: '1rem' }}>
+          <span>Changed your mind?</span>
+          <Button
+            style={{ color: 'white', marginLeft: '12px' }}
+            kind="ghost"
+            onClick={onUndoRowEdit}
+            size="small"
+            type="button"
+          >
+            Undo edits
+          </Button>
+        </div>
+      }
+      timeout={5000}
+      title="Your changes have been saved."
+    />
+  );
+
+  return (
+    <div>
+      {showToast ? myToast : null}
+      <MyTable
+        id="table"
+        secondaryTitle="My editable table"
+        view={{
+          toolbar: {
+            activeBar: showRowEditBar ? 'rowEdit' : undefined,
+            rowEditBarButtons: saveCancelButtons,
+          },
+          table: {
+            rowActions: rowActionsState,
+            singleRowEditButtons: saveCancelButtons,
+          },
+        }}
+        // WORKAROUND for #3406
+        // Ideally we would only ever use the currentData here, but the stateful table doesn't pick
+        // up isolated changes to view.toolbar.rowEditBarButtons or view.table.singleRowEditButtons so we
+        // have to trigger the TABLE_REGISTER action by some other means in order to render the changes
+        // made to the disabled state of the save button. That is why we then use rowEditData.
+        data={selectedTableType === 'StatefulTable' && !isPristine ? rowEditData : currentData}
+        actions={{
+          table: { onApplyRowAction },
+          toolbar: { onShowRowEdit: onShowMultiRowEdit },
+        }}
+        options={{
+          preserveColumnWidths: true,
+          hasResize: true,
+          hasRowEdit,
+          hasSingleRowEdit,
+          preserveCellWhiteSpace,
+          hasRowActions,
+        }}
+        columns={columns}
+      />
+    </div>
+  );
+};
+WithDataEditing.storyName = 'With data editing';
+WithDataEditing.parameters = {
+  component: Table,
+  docs: {
+    page: EditDataREADME,
   },
 };