+++ packages/react/src/constants/CardPropTypes.js
@@ -89,7 +89,7 @@ export const DashboardColumnsPropTypes = PropTypes.shape({
   xs: PropTypes.number,
 });
 
-export const ValueCardPropTypes = {
+export const ValueContentPropTypes = {
   content: PropTypes.shape({
     attributes: PropTypes.arrayOf(AttributePropTypes).isRequired,
   }),